Moves components into folder groups
